+++ modules/hanna-react/src/Datepicker.tsx
@@ -27,7 +27,7 @@ export type DatepickerProps = {
   minDate?: Date;
   maxDate?: Date;
   localeCode?: 'is' | 'en' | 'pl'; // default 'is', we can add more langs later...
-  dateFormat?: string;
+  dateFormat?: string | Array<string>;
   isStartDate?: boolean;
   isEndDate?: boolean;
   inputRef?: RefObject<HTMLInputElement>;
@@ -177,7 +177,18 @@ export const Datepicker = (props: DatepickerProps) => {
               selected={value}
               name={name}
               locale={localeCode}
-              dateFormat={dateFormat}
+              dateFormat={
+                // NOTE: Force all dateFormat values into Array<string> to temporarily work around
+                // a bug in the current version of react-datepicker where invalid **string** values
+                // are re-parsed with `new Date()`, causing surprising behavior in certain situations
+                // AND wheree Arrayed formats get parsed in order of "increasing priority".
+                // Revert back to the plain `dateFormat={dateFormat}` pass-through once this PR has been
+                // accepted and released: https://github.com/Hacker0x01/react-datepicker/pull/3903
+                typeof dateFormat === 'string'
+                  ? [dateFormat]
+                  : dateFormat.slice(0).reverse()
+                // dateFormat
+              }
               onChange={(date: Date | null) => {
                 onChange(date || undefined);
                 const inputElm = inputRef?.current;